refactor: 회원과 상품 연관관계 추가 (#29)
* refactor: 상품 등록, 수정 메소드 파라미터 변경

- requestDto -> entity 변경

* refactor: Controller 에 ResponseEntity 타입 명시

- ItemApiController: 상품 삭제 메소드 HttpStatus 타입 명시

* refactor: 회원, 상품간 연관관계 추가

- 회원과 상품은 일(회원)대다(상품) 관계이다

* test: 상품 등록, 수정 메소드 파라미터 변경에 대한 테스트 코드 수정

* test: 회원과 상품 연관관계 테스트

- 회원은 상품을 등록할 수 있다
- 상품에 권한을 가진 회원은 상품 정보를 수정할 수 있다
- 상품에 권한을 가진 회원은 상품 정보를 삭제할 수 있다

* refactor: 상품 생성 메소드 생성

* test: 회원과 상품 연관관계 도메인 테스트

* test: 회원과 상품 연관관계 도메인 테스트

* refactor: 회원과 상품 연관관계 비즈니스 로직 추가

* test: 회원과 상품 연관관계 비즈니스 로직 테스트

* refactor: 회원과 상품 연관관계 Controller 추가

- 상품 등록, 수정, 삭제 시 로그인을 확인한다
- 상품 수정, 삭제 시 상품에 대한 권한을 확인한다

* test: 회원과 상품 연관관계 API 테스트

- 상품 등록, 수정, 삭제 시 로그인을 확인한다
- 상품 수정, 삭제 시 상품에 대한 권한을 확인한다

* feat: 로그인 확인 중복 로직 Interceptor 처리

* refactor: 로그인 확인 중복 로직 제거

* refactor: HTTP Method GET 예외 허용

* test: 로그인 인터셉터 적용 후 API 테스트

- 로그인 여부 확인
- 게시글 권한 확인

+++ back/src/main/java/com/gagi/market/item/api/ItemApiController.java
@@ -1,11 +1,14 @@
 package com.gagi.market.item.api;
 
+import com.gagi.market.config.auth.LoginMember;
 import com.gagi.market.item.api.dto.ItemRequestDto;
 import com.gagi.market.item.api.dto.ItemResponseDto;
 import com.gagi.market.item.domain.Item;
 import com.gagi.market.item.service.ItemService;
+import com.gagi.market.member.api.dto.SessionMember;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -51,23 +54,33 @@ public class ItemApiController {
     }
 
     @PostMapping
-    public ResponseEntity<ItemResponseDto> createItem(@RequestBody ItemRequestDto requestDto) {
-        Item item = itemService.createItem(requestDto);
+    public ResponseEntity<ItemResponseDto> createItem(@LoginMember SessionMember member,
+                                                      @RequestBody ItemRequestDto requestDto) {
+        Item item = itemService.createItem(requestDto.toEntity(), member.getMemberEmail());
         return ResponseEntity
                 .created(URI.create(ITEM_API_URI + "/" + item.getItemId()))
                 .body(ItemResponseDto.of(item));
     }
 
     @PutMapping("/{itemId}")
-    public ResponseEntity<ItemResponseDto> updateItem(@PathVariable Long itemId, @RequestBody ItemRequestDto requestDto) {
-        Item item = itemService.updateItem(itemId, requestDto);
+    public ResponseEntity<ItemResponseDto> updateItem(@LoginMember SessionMember member,
+                                                      @PathVariable Long itemId,
+                                                      @RequestBody ItemRequestDto requestDto) {
+        if (!itemService.checkPermissionOfItem(itemId, member.getMemberEmail())) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
+        Item item = itemService.updateItem(itemId, requestDto.toEntity());
         return ResponseEntity
                 .created(URI.create(ITEM_API_URI + "/" + item.getItemId()))
                 .body(ItemResponseDto.of(item));
     }
 
     @DeleteMapping("/{itemId}")
-    public ResponseEntity deleteItem(@PathVariable Long itemId) {
+    public ResponseEntity<HttpStatus> deleteItem(@LoginMember SessionMember member,
+                                                 @PathVariable Long itemId) {
+        if (!itemService.checkPermissionOfItem(itemId, member.getMemberEmail())) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
         itemService.deleteItem(itemId);
         return ResponseEntity
                 .ok()
